zsh: check if pikaur is installed before trying to move cache dir

+++ zsh/move_pikaur_cache_in_ram.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
+if [ -z "$(command -v pikaur)" ]; then
+  exit
+fi
 
 rm -rf ~/.cache/pikaur/build
 mkdir -p /tmp/pikaur_cache
